requirements.txt
Django==1.6.7
South==1.0
colander==1.0
django-bootstrap3==4.11.0
django-localflavor==1.0
django-pyodbc==0.2.5
django-pyodbc-azure==1.2.1
django-selectable==0.9.0
django-smart-selects==1.0.9
iso8601==0.1.10
mechanize==0.2.5
nose==1.3.4
phonenumbers==7.0.1
pyodbc==3.0.7
pyserial==2.7
pysms==0.2
python-gsmmodem==0.9
python-messaging==0.5.12
six==1.8.0
smspdu==1.0
translationstring==1.3
virtualenv==1.9.1
virtualenvwrapper-win==1.1.5
xlrd==0.9.2
xlutils==1.7.0
xlwt==0.7.5
